support tarballs with directory structure
Rather than doing 'for x in *' to iterate through the files
that were in the tarball, we dump the tar file list to a file, then
read that file for a file list.

The benefit of this is that files in subdirectories are also found.
I chose this rather than doing something like 'find * -type f' just
to avoid the otherwise needless 'find'.

One note, this change was prompted by the format of eucalyptus's
tarball images, which look like
  $ tar -tzf ../euca-centos-5.3-x86_64.tar.gz 
  euca-centos-5.3-x86_64/
  euca-centos-5.3-x86_64/centos.5-3.x86-64.img
  euca-centos-5.3-x86_64/xen-kernel/
  euca-centos-5.3-x86_64/xen-kernel/vmlinuz-2.6.27.21-0.1-xen
  euca-centos-5.3-x86_64/xen-kernel/initrd-2.6.27.21-0.1-xen
  euca-centos-5.3-x86_64/kvm-kernel/
  euca-centos-5.3-x86_64/kvm-kernel/vmlinuz-2.6.28-11-generic
  euca-centos-5.3-x86_64/kvm-kernel/initrd.img-2.6.28-11-generic

However, because 'xen-kernel' is found first in the list, the xen kernel
and ramdisk will be selected.  There is no way to specify that you
want the kvm kernel.